Update ServiceProvider for multiple instance support
The changes in ServiceProvider now allow for multiple instances of singleton and transient services by storing them in lists. Added GetServices methods to retrieve all single or multiple instances of a specific service. Updated the generation of services to append new instances to the list of either singleton or transient services instead of replacing the previous value.